requirements.txt
discord.py==2.3.1
numpy==1.24.3
Pillow==9.5.0
Pillow==9.5.0
python-dotenv==1.0.0
Requests==2.31.0
